requirements.txt
nose2==0.9.1
python-dotenv==0.12.0
SQLAlchemy==1.3.13
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
passlib==1.7.2
flask-jwt-extended==3.24.1
flask_login==0.5.0
mysqlclient==1.4.6
kubernetes==11.0.0
flask-socketio==4.3.0
eventlet==0.25.2
requests-toolbelt==0.9.1
python-ldap==3.3.0
python3-saml==1.9.0
plotly==5.3.0
PyJWT==1.7.1
python-socketio==4.5.1
python-engineio==3.12.1
psutil==5.6.3
graphviz
pycryptodome==3.9.8
pandas==1.3.0